backend/requirements.txt

# Django Framework
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0

# AI/ML Libraries - Core
transformers==4.30.0
torch==2.0.0
sentence-transformers==2.2.2
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.2.2

# Vector Search & FAISS
faiss-cpu==1.7.4

# ✅ NEW: Speech-to-Text (Whisper)
faster-whisper>=0.10.0

# ✅ NEW: Audio Processing Dependencies
av>=10.0.0
tokenizers>=0.13.0
huggingface-hub>=0.16.0

# ✅ OPTIONAL: Enhanced Audio Processing
# Uncomment if you need advanced audio features
# librosa>=0.10.0
# soundfile>=0.12.0
# pydub>=0.25.0

# HTTP Requests (for Gemini API)
requests==2.31.0

# Utilities
python-dotenv==1.0.0
Pillow==9.5.0

# ✅ NEW: Logging & Monitoring
colorama>=0.4.6

# ✅ NEW: Unicode & Text Processing
unidecode>=1.3.6
regex>=2023.0.0
